content-service/requirements.txt
flask==2.0.1
pymongo==3.12.0
redis==3.5.3
python-dotenv==0.19.0
flask-cors==2.1.3
gunicorn==20.1.0
marshmallow==3.13.0
requests==2.26.0 